chore: Add Prisma database client
A new file `prismadb.ts` is added to the `libs` directory. This file contains the code to create a new instance of the PrismaClient and assign it to the global object. This ensures that the Prisma instance is only created once. The purpose of this change is to integrate Prisma into the project for database operations.