refactor: Update UserMenu component to accept currentUser prop
The UserMenu component in the UserMenu.tsx file is updated to accept a currentUser prop of type User | null. This prop is used to conditionally render different menu items based on whether the user is logged in or not. This change improves the flexibility and reusability of the UserMenu component.