Logo component updated to navigate to the homw when click
The Logo component in the Logo.tsx file is updated to include lazy loading for images. This change improves the page load performance by loading the logo image only when it is visible on the screen. Additionally, a click event is added to navigate to the home page when the logo is clicked.